# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: IncomeExpense API
  version: 1.0.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AuthResponse:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        token: { type: string }
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
            example: { name: "Kayahan", email: "kayahan@example.com", password: "P@ssw0rd!" }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                remember: { type: boolean }
            example: { email: "kayahan@example.com", password: "P@ssw0rd!" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      accessToken: { type: string }
                      user:
                        type: object
                        properties:
                          id: { type: string }
                          email: { type: string }
                          name: { type: string }
                  - type: object
                    properties:
                      mfa_required: { type: boolean }
                      tmpToken: { type: string }
  /transactions:
    get:
      summary: List my transactions
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK